+++ common/src/main/java/brooklyn/networking/common/subnet/PortForwarderAsyncImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package brooklyn.networking.subnet;
+package brooklyn.networking.common.subnet;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
