+++ portforwarding/src/main/java/brooklyn/networking/portforwarding/PortForwarderIptables.java
@@ -34,7 +34,7 @@ import brooklyn.location.access.PortForwardManager;
 import brooklyn.location.basic.PortRanges;
 import brooklyn.location.basic.SshMachineLocation;
 import brooklyn.management.ManagementContext;
-import brooklyn.networking.subnet.PortForwarder;
+import brooklyn.networking.common.subnet.PortForwarder;
 import brooklyn.util.net.Cidr;
 import brooklyn.util.net.HasNetworkAddresses;
 import brooklyn.util.net.Protocol;
