+++ portforwarding/src/test/java/brooklyn/networking/portforwarding/DockerSubnetTierLiveTest.java
@@ -29,7 +29,6 @@ import brooklyn.entity.basic.EntityAndAttribute;
 import brooklyn.entity.proxying.EntitySpec;
 import brooklyn.event.AttributeSensor;
 import brooklyn.event.basic.BasicAttributeSensor;
-import brooklyn.networking.portforwarding.DockerPortForwarder;
 import brooklyn.networking.subnet.SubnetTier;
 import brooklyn.test.Asserts;
 import brooklyn.util.net.Cidr;
